refactored code, rewrote unit_tests based on the new Python package setup
