+++ bin/stlparse
@@ -1,0 +1,16 @@
+#!/bin/bash
+
+# store the directory where user invoked the stl command
+exec_dir="$(pwd)"
+
+# go to the directory where the STL Script Interpreter is located
+cd $(dirname $0)/../
+
+if [ $# -eq 0 ]; then
+    # no command-line arguments, REPL mode
+    python3 -m src.parser
+else
+    # with 1 or more command-line arguments, program mode
+    python3 -m src.parser $exec_dir/$*
+fi
+
