+++ src/AST_Collection/STL_expr.py
@@ -3,10 +3,10 @@
 from sys import stdout, path
 path.append("..") # Adds higher directory to python modules path.
 
-from tools import String_Builder
+from src.tools import String_Builder
 
-from core_AST import STL_Expr
-from val import Meta_Id_Val
+from src.AST_Collection.core_AST import STL_Expr
+from src.AST_Collection.val import Meta_Id_Val
