+++ src/AST_Collection/stmt.py
@@ -5,11 +5,11 @@
 from sys import stdout, path
 path.append("..") # Adds higher directory to python modules path.
 
-from tools import String_Builder
+from src.tools import String_Builder
 
-from core_AST import Stmt
-from val import Boolean_Val
-import exceptions
+from src.AST_Collection.core_AST import Stmt
+from src.AST_Collection.val import Boolean_Val
+import src.exceptions as exceptions
 
 class Block_Stmt(Stmt):
     def __init__(self, stmt_list):
