--- src/interp
@@ -1,2 +1,0 @@
-#!/bin/bash
-python3 main.py $*
