--- src/lexer
@@ -1,1 +1,0 @@
-python3 lexer.py $*
