+++ src/test_suite/normal_assign.stl
@@ -10,17 +10,20 @@ val j_2: String = "hello"
 val k_2: Float = 3.1415926
 val l_2: Boolean = false
 
+
 // untyped var decl
 var i_3 = 10
 var j_3 = "hello"
 var k_3 = 3.1415926
 var l_3 = false
 
+
 // typed var decl
 var i_4: Int = 10
 var j_4: String = "hello"
 var k_4: Float = 3.1415926
 var l_4: Boolean = false
+
 
 println i_1
 println j_1
@@ -61,6 +64,8 @@ println i_4
 println j_4
 println k_4
 println l_4
+
+
 
 /*
