--- src/utest
@@ -1,2 +1,0 @@
-#!/bin/bash
-python3 unit_tests.py
